inching closer to world domination
This chunk-o-work obsoletes the old ColumnCell stateless renderers
and dependency on a ColumnController (which will eventually be
moved internally to the the list layout).

Instead I introduce the DataViewChild, which is more like a widget:

  - it holds a real allocation
  - it holds its bound model object
  - it can hold arbitrary internal state

Since these layout children hold real state, awful hacks need not
be employed for doing things like interaction or even animation.

Similarly, awful hacks are not needed at the view interaction level
to proxy proper input events to the layout children. This greatly
simplifies the interaction code. Also, since a layout child knows
its allocation at all times, it can request to be redrawn, so this
is less work the view needs to do.

ColumnCellAlbum has been ported to the new DataViewChild API. Next
up will be animating this beast. Yes.

Finally, DataViewLayoutGrid has a new ChildAllocator property that
must be assigned to. This is a function that returns an instance
of a DataViewChild. Alternatively, if subclassing DataViewLayoutGrid,
the virtual CreateChild method may be overridden. This eliminates
the need for AlbumListView to provide a subclassed DataViewLayoutGrid
just to specify its children:

 +            ViewLayout = new DataViewLayoutGrid () {
 +                ChildAllocator = () => new DataViewChildAlbum (),
 +                View = this
 +            };

Keep cool my babies!