hack to work around endless loop
If a column size is ever computed to be 0, which is possible in the rare
case where only one flexible column will be visible, and sizing is just
right for computing the flexible scale value to be 1/col, later in the
column size calculations, this 0 value will cause other values to become
NaN, which will then result in a lovely infinite loop.

When applying the scale back, we arrive at this dreaded 0 value.

This entire section of code needs some serious review as there's much
that is not clearly understood here, but hey, it works (99.9% of the time
at least), so I'm nervous about "fixing" it.