[Hyena.Data.Sqlite] Load result data into array
For the same reason we did before with HyenaSqliteArrayDataReader -- we
put all the result rows' data into managed memory so the sqlite
statement can be finalized.  Ideally we wouldn't have to do this, but
for now it's the path of least resistance.  This is only done for
HyenaSqliteCommand queries, not for queries done directly with the new
sqlite binding.