ListView: Fix a cell rendering glitch (bgo#636057)
ColumnCell x-coordinate was being incorrectly calculated in order to call
QueueDirtyRegion in response to mouse events. The original calculation used
ListView's Allocation, which doesn't account for the offset of
TotalBorderWidth. This glitch was mostly noticeable with ColumnCellCheckBox
items. Allocation has been change to list_rendering_alloc to correct the
calculation.

Signed-off-by: Gabriel Burt <gabriel.burt@gmail.com>