QueryField: check for NULL before LIKE
When getting the SQL result of crossing a QueryField
with a QueryValue, the check for NULL was being printed
after the rest of the query.

In theory, the check for NULL should be done before so
you really gain the performance of only running one part
of the query in case the first part is already TRUE (OR
operator) or FALSE (AND operator). Maybe SQLite was
already smart enough to optimize this regardless, but
the query is anyway clearer now.

It is now also more in line with the .NET API:
String.IsNullOrEmpty()