Query: fix FileLocation queries for the StartsWith operator
In the same way user queries were already being escaped [1] so
the user didn't need to do it manually (i.e. when creating a
Smart Playlist in Banshee), hyena should be smart enough to add
the file scheme to the value queried in case the operator
StartsWith is used, so the Uri is correctly formed in the LIKE
filter. Fixes bgo#644145.

[1] http://mail.gnome.org/archives/commits-list/2010-March/msg02864.html