ListView: Do not enable a11y when it would cause freezes (bgo#696111)
Recent versions of glib or gobject (2.35.4 and above) don't allow
anymore to add GInterfaces after class initialization, which causes the
Mono binding to fail when implementing interfaces in managed land (filed
in https://bugzilla.xamarin.com/show_bug.cgi?id=11510).

We're  affected by this particularly when implementing the AtkTable
interface for the accessibility support of the ListView custom widget.
The consequence would be that the class would not implement the
interface properly, which would cause lots of exceptions in
OnCreateAccessible signals, high CPU usage and make the app hang.

The workaround for now is create a dummy instance of ListViewAccessible
class, and not enable accessibility if any exception is caught at
creation time.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>